Create a folder for frontend when a department is created
Overridden the default save method for the Department model such that
when a department is created, a folder is created inside the root
folder with the short code of the department and a dummy.html file is
copied into it with the name index.html. Add required contents in the
dummy.html file.